--- address.dart
@@ -1,70 +1,0 @@
-import 'package:flutter/material.dart';
-import 'newAddress.dart';
-import 'main.dart';
-
-class AddressScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        toolbarHeight: 80, 
-        title: Row(
-          children: [
-            Icon(Icons.location_on, color: Colors.red, size: 35,),
-            SizedBox(width: 8),
-            Text(
-              'Address',
-              style: TextStyle(color: Colors.red, fontSize: 24, fontWeight: FontWeight.bold),
-            ),
-          ],
-        ),
-        automaticallyImplyLeading: false,
-        actions: [
-          Container(
-            margin: EdgeInsets.symmetric(horizontal: 25.0),
-            child: IconButton(
-              icon: Icon(Icons.chevron_left, color: Colors.red, size: 35),
-              onPressed: () {
-                Navigator.pushReplacement(
-                  context,
-                  MaterialPageRoute(builder: (context) => MyApp()),
-                );
-              },
-            ),
-          ),
-        ],
-      ),
-      body: Center(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(context,
-                  MaterialPageRoute(builder: (context) => NewAddress()),
-                );
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red, 
-                foregroundColor: Colors.white, 
-                padding: EdgeInsets.symmetric(horizontal: 20, vertical: 12),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(10), 
-                ),
-              ),
-              child: Row(
-                mainAxisSize: MainAxisSize.min, 
-                children: [
-                  Icon(Icons.add, size: 20, color: Colors.white), 
-                  SizedBox(width: 8), 
-                  Text('Add Address'),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
