+++ bazaroo_client/lib/startupscreen.dart
