--- main.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'startupscreen.dart';
-
-void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: StartupScreen(), 
-    );
-  }
-}
